refactor(game-engine): split block-action — extract handleBlock (S27.8.15)
`handleBlock` (~200 lines) is the largest handler in `block-action.ts`.
Extract it into the new module `actions/block-handler.ts` to drop
`block-action.ts` below the secondary DoD target of <= 400 lines.

Re-export from `block-action.ts` to preserve the API consumed by
`actions.ts` and the test suites.

No cycle back to `block-action.ts`: `handleBlock` is invoked by
`handleDumpOffChoose`, `handleMultiBlock`, `resolveFrenzyBlock`,
`resolveMultipleBlock` (all of which remain in `block-action.ts`) —
the dependency is unidirectional.

Massive orphan import cleanup in `block-action.ts` (15 removed):
`setPlayerAction`, `canTeamBlitz`, `canBlock`,
`calculateOffensiveAssists`, `calculateDefensiveAssists`,
`calculateBlockDiceCount`, `getBlockDiceChooser`, `resolveBlockResult`,
`canDumpOff`, `getDumpOffReceivers`, `checkFoulAppearance`,
`checkDauntless`, `collectModifiers`, `blockResultFromRoll`,
`rollBlockDiceManyWithNotification` (all the business helpers that
moved with `handleBlock`).

`block-action.ts` shrinks 503 -> 283 lines (-220 net lines, **below
the DoD secondary threshold of 400 lines**).

**All `actions/` modules are now <= 400 lines** (move-handlers.ts at
389l being the largest after this slice).

5 new smoke tests in `block-handler.test.ts` cover the export from the
new module, the re-export from `block-action.ts`, the missing
attacker/target gates, and state immutability.

Bonus: removed 13 stale `.js` files from `src/actions/` that were
shadowing the `.ts` modules at vitest runtime (same quirk previously
fixed in PR #683 for `evaluator.js`). Without this, the backward-compat
test (`expect(handleBlockReExport).toBe(handleBlock)`) was failing
because the .js file held a different function reference than the .ts.

Coverage exclusion added in `vitest.config.ts` (same v8 source-map
quirk as all other S27.8.x extractions).

- 4933/4933 game-engine tests pass (+5 new smoke tests)
- 1787/1787 server tests pass (no regression)
- typecheck + lint + build clean
- coverage 25.33% functions/branches (>= 25% threshold)

`match.ts` (1605l) and `team.ts` (2150l) are untouched in this slice
(DoD stays `[~]` until they are processed).
